+++ src/BrregAPiServiceProvider.php
@@ -4,22 +4,20 @@ namespace Tor2r\BrregAPi;
 
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
-use Tor2r\BrregAPi\Commands\BrregAPiCommand;
 
 class BrregAPiServiceProvider extends PackageServiceProvider
 {
     public function configurePackage(Package $package): void
     {
-        /*
-         * This class is a Package Service Provider
-         *
-         * More info: https://github.com/spatie/laravel-package-tools
-         */
         $package
             ->name('laravel-brreg-api')
-            ->hasConfigFile()
-            ->hasViews()
-            ->hasMigration('create_laravel_brreg_api_table')
-            ->hasCommand(BrregAPiCommand::class);
+            ->hasConfigFile();
+    }
+
+    public function packageRegistered(): void
+    {
+        $this->app->singleton(BrregAPi::class, function () {
+            return new BrregAPi;
+        });
     }
 }