+++ src/Facades/BrregAPi.php
@@ -5,6 +5,11 @@ namespace Tor2r\BrregAPi\Facades;
 use Illuminate\Support\Facades\Facade;
 
 /**
+ * @method static array getByOrgnr(string $orgnr)
+ * @method static array searchByOrgnr(string $orgnr, ?int $limit = null)
+ * @method static array searchByName(string $name, ?int $limit = null)
+ * @method static \Tor2r\BrregAPi\BrregAPi voluntary()
+ *
  * @see \Tor2r\BrregAPi\BrregAPi
  */
 class BrregAPi extends Facade